requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
ckzg==1.0.1
click==8.1.7
colorama==0.4.6
cytoolz==0.12.3
eth-abi==5.1.0
eth-account==0.12.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==4.2.2
eth-utils==4.1.0
flask==3.0.3
frozenlist==1.4.1
gunicorn==22.0.0
hexbytes==0.3.1
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
MarkupSafe==2.1.5
multidict==6.0.5
packaging==23.2
parsimonious==0.10.0
pkgutil-resolve-name==1.3.10
protobuf==5.26.1
py-solc-x==2.0.2
pycryptodome==3.20.0
pydantic==2.7.1
pydantic-core==2.18.2
pyunormalize==15.1.0
pywin32==306
referencing==0.35.0
regex==2024.4.28
requests==2.31.0
rlp==4.0.1
rpds-py==0.18.0
toolz==0.12.1
typing-extensions==4.11.0
urllib3==2.2.1
web3==6.18.0
websockets==12.0
werkzeug==3.0.2
yarl==1.9.4
zipp==3.18.1
